Won't try to remove the empty list message

+++ app/src/main/java/edu/umich/globalchallenges/thirdeye/FileViewerFragment.java
@@ -196,12 +196,14 @@ public class FileViewerFragment extends Fragment implements FlexibleAdapter.OnIt
     @Override
     public void onItemLongClick(int position) {
         toDelete = databaseMessage.get(position).get(1); // Grab filename of thing we want deleted
-        Snackbar deletebar = Snackbar.make(getActivity().getWindow().getDecorView().findViewById(android.R.id.content), "Delete file on server?", Snackbar.LENGTH_LONG);
-        deletebar.getView().setBackgroundColor(getResources().getColor(R.color.colorPrimary));
-        TextView messagetext = (TextView) deletebar.getView().findViewById(android.support.design.R.id.snackbar_text);
-        messagetext.setTextColor(Color.BLACK);
-        deletebar.setAction("Yes", new FileViewerFragment.DeleteListener());
-        deletebar.show();
+        if(!toDelete.contentEquals("Try refreshing")) { // Don't try to remove the empty list message
+            Snackbar deletebar = Snackbar.make(getActivity().getWindow().getDecorView().findViewById(android.R.id.content), "Delete file on server?", Snackbar.LENGTH_LONG);
+            deletebar.getView().setBackgroundColor(getResources().getColor(R.color.colorPrimary));
+            TextView messagetext = (TextView) deletebar.getView().findViewById(android.support.design.R.id.snackbar_text);
+            messagetext.setTextColor(Color.BLACK);
+            deletebar.setAction("Yes", new FileViewerFragment.DeleteListener());
+            deletebar.show();
+        }
     }
 
     /**
